Implementations of the suggestions
I have changed the baudrate calculation so that we can enter 9600 and it automatically calculates it to 103.

I have also added the unsigned int before baudrate. I was wondering why it was giving an error lol.

as for the setup/init pin initialization, i removed setting all pins to digital and just kept setting the pins we need (_TRISB6 = 0; _TRISB10 = 1;)

as for back &= 63, that was originally copied from the discussion to implement a circular buffer. however, it didn't work when i tried implementing it in my lab so i have upddated it to a version that works and is self explanatory.

Finally, I have added a string getter. It requires a pointer string  and also the expected size of the string.